refactor(ui): extrair utilitários de UI para Craftimizer.UI compartilhado
Cria Craftimizer.UI como biblioteca standalone (sem Dalamud) com todas as
primitivas de ImGui reutilizáveis. O plugin agora referencia Craftimizer.UI
em vez de definir essas classes internamente.

- Novo projeto Craftimizer.UI com ImGuiUtils, ImRaii2, ProgressBarComponent,
  Theme, Colors, IUiServices/UiServices, ProgressBarType, UIConstants
- DalamudUiServices implementa IUiServices para bridge de tipos Dalamud→ImGui.NET
- Plugin renomeia classes conflitantes: ImRaii2→ImRaiiPlot, ProgressBarComponent.Solver→SolverProgressBar
- Call sites corrigidos: FontAwesomeIcon→(int), ImTextureID.Handle→ulong overload,
  ImGuiInputTextFlags/ButtonFlags→int, ImFontPtr via unsafe cast ou UiServices
- GlobalUsings.cs e IFontHandleExtensions.cs adicionados ao plugin

Version: 2.18.1.1

+++ Craftimizer/Windows/CosmicTracker.cs
@@ -156,7 +156,7 @@ public sealed class CosmicTracker : Window, IDisposable
         if (_progress is not { } p)
         {
             ImGuiUtils.DrawEmptyState(
-                FontAwesomeIcon.Star,
+                (int)FontAwesomeIcon.Star,
                 "Sem dados ainda",
                 "Entre em qualquer mapa de Cosmic Exploration.");
             return;
@@ -187,7 +187,7 @@ public sealed class CosmicTracker : Window, IDisposable
 
         if (!anyRendered)
             ImGuiUtils.DrawEmptyState(
-                FontAwesomeIcon.CheckCircle,
+                (int)FontAwesomeIcon.CheckCircle,
                 "Tudo concluído",
                 "Todos os tipos de pesquisa foram completados.\nDesative o filtro para visualizá-los.");
